Add new image to run rt with cyclictest
Create an image to run cyclictest, and
modify scripts on hack to use it.

Use different cyclictest image

Remove rt-tests image and start using new cyclictest
image. This needs to be combined with some stress on
the system, to allow proper measure.

Use different cyclictest image

Remove rt-tests image and start using new cyclictest
image. This needs to be combined with some stress on
the system, to allow proper measure.

Add new podman test to use cyclictest and stress-ng